# api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: floTeam@gmail.com
  description: api for flow learning
  termsOfService: http://flowLearning.com/terms/
  title: Flow Learning - OpenAPI 3.1
  version: 1.0.0
servers:
- url: https://floLearning.com/api/v1
...
paths:
...
  /users/auth:
    get:
      description: return user
      operationId: AuthUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser_200_response'
          description: students matching criteria
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser_404_response'
          description: No user found with that data
      summary: return authenticated user details
      tags:
      - user
  /user/login:
    post:
      description: Login using SSO
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginUser_request'
        description: Login request
        required: true
      responses:
        "200":
          description: Login successful
          headers:
            X-Auth-Token:
              description: Authentication token
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: SSO provider
        "401":
          description: Authentication failed
      summary: Login to the system or create account
      tags:
      - user
  /user/logout:
    post:
      description: Revokes the current user's authentication session.
      operationId: logoutUser
      responses:
        "200":
          description: User logged out successfully
        "401":
          description: Unauthorized access
      summary: Log the user out of the system
      tags:
      - user
  /user:
    put:
      description: This can only be done by the logged in user.
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateUser_request'
        description: Update an existent user in the store
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
      summary: Update user
      tags:
      - user
  /user/{userId}:
    delete:
      description: This can only be done by the logged in user or admin.
      operationId: deleteUser
      parameters:
      - description: The name that needs to be deleted
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: User Deleted
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      summary: Delete user
      tags:
      - user
    get:
      description: ""
      operationId: getUserByName
      parameters:
      - description: 'The name that needs to be fetched. Use user1 for testing. '
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/updateUser_request'
          description: successful operation
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      summary: Get user by user name
      tags:
      - user
components:
  schemas:
...
    addTopic_200_response_nodeData_youtubeLinks_inner_addedBy:
      example:
        id: dkd94njd
        username: super123
      properties:
        username:
          example: super123
          type: string
        id:
          example: dkd94njd
          type: string
    addTopic_200_response_nodeData_youtubeLinks_inner:
      example:
        addedBy:
          id: dkd94njd
          username: super123
        link: https://youtu.be/1MKKK94eGUo?si=uRr8UmZ0sgNl4BA0&t=397
        votes: 21
        dateAdded: 2024-08-19T19:25:42.568Z
      properties:
        link:
          example: https://youtu.be/1MKKK94eGUo?si=uRr8UmZ0sgNl4BA0&t=397
          format: url
          type: string
        votes:
          example: 21
          format: int32
          type: integer
        addedBy:
          $ref: '#/components/schemas/addTopic_200_response_nodeData_youtubeLinks_inner_addedBy'
        dateAdded:
          example: 2024-08-19T19:25:42.568Z
          format: date-time
          type: string
...
    AuthUser_200_response:
      example:
        isAuth: true
        role: 0
      properties:
        isAuth:
          type: boolean
        role:
          example: 0
          type: integer
      required:
      - isAdmin
      - isAuth
      - role
    AuthUser_404_response:
      example:
        isAuth: "False"
        error: true
      properties:
        isAuth:
          example: "False"
          type: boolean
        error:
          type: boolean
    loginUser_request:
      properties:
        ssoProvider:
          enum:
          - google
          - facebook
          - microsoft
          - x
          - apple
          - discord
          type: string
    updateUser_request_battleTestedUp_inner:
      example:
        topic: bjj
        title: armbar
        nodeId: 2024-12-09T04:10:00.350Z
      properties:
        topic:
          example: bjj
          type: string
        title:
          example: armbar
          type: string
        nodeId:
          example: 2024-12-09T04:10:00.350Z
          format: date-time
          type: string
    updateUser_request:
      example:
        freshUp:
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        lastLogin: 2000-01-23T04:56:07.000+00:00
        lastName: James
        role: 1
        edited:
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        created:
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        freshDown:
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        reputation: 34
        description: engineer
        videoUp:
        - www.youtube.com
        - www.youtube.com
        battleTestedUp:
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        isFlagged: true
        createdAt: 2000-01-23T04:56:07.000+00:00
        firstName: John
        provider: google
        videoDown:
        - www.youtube.com
        - www.youtube.com
        battleTestedDown:
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        - topic: bjj
          title: armbar
          nodeId: 2024-12-09T04:10:00.350Z
        location: Austin TX
        id: 2451632
        email: john@email.com
        updatedAt: 2000-01-23T04:56:07.000+00:00
        username: theUser
        linked:
        - addedBy:
            id: dkd94njd
            username: super123
          link: https://youtu.be/1MKKK94eGUo?si=uRr8UmZ0sgNl4BA0&t=397
          votes: 21
          dateAdded: 2024-08-19T19:25:42.568Z
        - addedBy:
            id: dkd94njd
            username: super123
          link: https://youtu.be/1MKKK94eGUo?si=uRr8UmZ0sgNl4BA0&t=397
          votes: 21
          dateAdded: 2024-08-19T19:25:42.568Z
      properties:
        provider:
          example: google
          type: string
        id:
          example: 2451632
          type: string
        lastLogin:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        createdAt:
          format: date-time
          type: string
        username:
          example: theUser
          type: string
        firstName:
          example: John
          type: string
        lastName:
          example: James
          type: string
        email:
          example: john@email.com
          type: string
        role:
          description: User Status
          example: 1
          format: int32
          type: integer
        reputation:
          example: 34
          type: integer
        description:
          example: engineer
          type: string
        location:
          example: Austin TX
          type: string
        isFlagged:
          type: boolean
        battleTestedUp:
          items:
            $ref: '#/components/schemas/updateUser_request_battleTestedUp_inner'
          type: array
        battleTestedDown:
          items:
            $ref: '#/components/schemas/updateUser_request_battleTestedUp_inner'
          type: array
        freshUp:
          items:
            $ref: '#/components/schemas/updateUser_request_battleTestedUp_inner'
          type: array
        freshDown:
          items:
            $ref: '#/components/schemas/updateUser_request_battleTestedUp_inner'
          type: array
        edited:
          items:
            $ref: '#/components/schemas/updateUser_request_battleTestedUp_inner'
          type: array
        created:
          items:
            $ref: '#/components/schemas/updateUser_request_battleTestedUp_inner'
          type: array
        linked:
          items:
            $ref: '#/components/schemas/addTopic_200_response_nodeData_youtubeLinks_inner'
          type: array
        videoUp:
          items:
            example: www.youtube.com
            type: string
          type: array
        videoDown:
          items:
            example: www.youtube.com
            type: string
          type: array